친구 추가 기능 수정

+++ daerkoob_back/src/main/java/com/project/daerkoob/model/TransferReview.java
@@ -13,7 +13,7 @@ public class TransferReview {
     Long id;
     String content;
     Long thumbCount;
-    Long score;
+    Double score;
     LocalDate registerDate;
     boolean thumbJudge;
 }
